# Improve performance using iterators.