This PR implements the main loop using iterators instead of indexing, and it also packs the precomputed sines and cosines into a single vec. It gives about a 10-20% speedup, depending on size.

```
Powers of 2/Real/64     time:   [226.65 ns 229.79 ns 235.53 ns]                                
                        change: [-14.022% -13.236% -12.139%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
Powers of 2/Real/128    time:   [441.92 ns 444.40 ns 447.74 ns]                                 
                        change: [-19.340% -16.507% -14.210%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  5 (5.00%) high mild
  11 (11.00%) high severe
Powers of 2/Real/256    time:   [958.09 ns 961.15 ns 964.68 ns]                                  
                        change: [-20.586% -15.971% -12.927%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 20 outliers among 100 measurements (20.00%)
  5 (5.00%) high mild
  15 (15.00%) high severe
Powers of 2/Real/512    time:   [2.0550 us 2.0629 us 2.0719 us]                                  
                        change: [-19.063% -15.949% -13.616%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  4 (4.00%) high mild
  14 (14.00%) high severe
Powers of 2/Real/4096   time:   [22.000 us 22.128 us 22.316 us]                                   
                        change: [-14.502% -10.812% -8.5181%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 20 outliers among 100 measurements (20.00%)
  5 (5.00%) high mild
  15 (15.00%) high severe
Powers of 2/Real/65536  time:   [460.05 us 464.23 us 469.37 us]                                   
                        change: [-24.418% -18.660% -14.209%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe

Powers of 7/Real/686    time:   [5.7136 us 5.7298 us 5.7478 us]                                  
                        change: [-14.803% -9.5756% -6.2845%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  3 (3.00%) high mild
  14 (14.00%) high severe
Powers of 7/Real/4802   time:   [53.528 us 53.680 us 53.845 us]                                  
                        change: [-15.611% -12.973% -10.657%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  2 (2.00%) high mild
  15 (15.00%) high severe
Powers of 7/Real/33614  time:   [485.49 us 489.09 us 494.81 us]                                   
                        change: [-14.698% -11.777% -9.2727%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe
```