@HEnquist 2020-07-22
Great! Thanks for this PR! I just ran the benchmarks and I see the same numbers. I was planning to do something about that indexed loop at some point, but I didn't get around to it yet because I didn't expect such a large improvement. I'll merge and publish a new version.