@HEnquist 2020-07-22
Done! Version 0.2.1 is available on crates.io.